Module dasi::model

Struct and enum definitions of values in the Discord model.

Reexports

pub use self::permissions::Permissions;

Modules

permissions

Structs

Attachment
File upload attached to a message

ChannelId
An identifier for a Channel

ChannelOverride
A channel-specific notification settings override

CurrentUser
Information about the logged-in user

Emoji
A parter custom emoji

EmojiId
An identifier for an Emoji

FriendSourceFlags
Flags for who may add this user as a friend.

Game
Information about a game being played

Invite
Information about an invite

LiveServer
Live server information

Member
Information about a member of a server

Message
Message transmitted over a text channel

MessageId
An identifier for a Message

Presence
A members's online status

PrivateChannel
Private text channel to another user

PublicChannel
Public voice or text channel within a server

ReadState
Summary of messages since last login

ReadyEvent
The "Ready" event, containing initial state

Relationship
Information on a friendship relationship this user has with another.

RichInvite
Detailed information about an invite, available to server managers

Role
Information about a role

RoleId
An identifier for a Role

Server
Static information about a server

ServerId
An identifier for a Server

ServerInfo
The basic information about a server only

Tutorial
Progress through the Discord tutorial

User
Broadly-applicable user information

UserId
An identifier for a User

UserServerSettings
User settings which influence per-server notification behavior

UserSettings
User settings usually used to influence client behavior

VoiceRegion
Information about an available voice region

VoiceState
A member's state within a voice channel

Enums

Channel
A private or public channel

ChannelType
The type of a channel

Event
Event received over a websocket connection

GameType
A type of game being played.

NotificationLevel
A user's online presence status

OnlineStatus
A user's online presence status

PermissionOverwrite
A channel-specific permission overwrite for a role or member

PossibleServer
A server which may be unavailable

RelationshipType
A type of relationship this user has with another.

VerificationLevel
A condition that new users must satisfy before posting in text channels