Module dasi::model::permissions

Structs

Permissions
Set of permissions assignable to a Role or PermissionOverwrite

Constants

ATTACH_FILES

BAN_MEMBERS

CHANGE_NICKNAMES
Change their own nickname

CREATE_INVITE

EMBED_LINKS

KICK_MEMBERS

MANAGE_CHANNELS
Create channels or edit existing ones

MANAGE_MESSAGES
Delete or edit messages by other users

MANAGE_NICKNAMES
Change the nickname of other users

MANAGE_ROLES
Implies all permissions

MANAGE_SERVER
Change the server's name or move regions

MENTION_EVERYONE
Trigger a push notification for an entire channel with "@everyone"

READ_HISTORY

READ_MESSAGES

SEND_MESSAGES

SEND_TTS_MESSAGES
Send text-to-speech messages to those focused on the channel

VOICE_CONNECT

VOICE_DEAFEN_MEMBERS

VOICE_MOVE_MEMBERS
Move users out of this channel into another

VOICE_MUTE_MEMBERS

VOICE_SPEAK

VOICE_USE_VAD
When denied, members must use push-to-talk